backend/requirements.txt
alembic==1.12.0
blinker==1.6.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
flask-swagger-ui==4.11.1
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
python-decouple==3.8
python-dotenv==1.0.1
SQLAlchemy==2.0.21
typing_extensions>=4.11.0
Werkzeug==3.0.0
numpy==1.26.4
pandas==2.2.2
Flask-Mail==0.10.0
openai>=2.6.1
pytest>=8.4.2
pytest-flask>=1.3.0
